Make 'Payload' abstract and break down 'errors'
Introduce a 'DataPayload' for payloads containing a primary data. Let it
hold the 'included' resources.

Introduce a 'MetaPayload' as it is possible to create a payload
essentially containing meta information.

Remove 'errors' from 'Payload' and create a specific payload subclass.
This is necessary because 'errors' cannot coexist with 'data'.

Remove the 'Payload.from_jsonapi' and 'Payload.from_errors' constructors
as they now have specific subclasses.